complete download issue
